+++ TelegramBotAPI-extensions-api/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/extensions/api/send/media/SendPhoto.kt
@@ -1,8 +1,7 @@
 package com.github.insanusmokrassar.TelegramBotAPI.extensions.api.send.media
 
 import com.github.insanusmokrassar.TelegramBotAPI.bot.RequestsExecutor
-import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.FileId
-import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.MultipartFile
+import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.*
 import com.github.insanusmokrassar.TelegramBotAPI.requests.send.media.SendPhoto
 import com.github.insanusmokrassar.TelegramBotAPI.types.ChatIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.MessageIdentifier
@@ -13,7 +12,7 @@ import com.github.insanusmokrassar.TelegramBotAPI.types.files.*
 
 suspend fun RequestsExecutor.sendPhoto(
     chatId: ChatIdentifier,
-    fileId: FileId,
+    fileId: InputFile,
     caption: String? = null,
     parseMode: ParseMode? = null,
     disableNotification: Boolean = false,
@@ -29,108 +28,4 @@ suspend fun RequestsExecutor.sendPhoto(
         replyToMessageId,
         replyMarkup
     )
-)
-
-suspend fun RequestsExecutor.sendPhoto(
-    chatId: ChatIdentifier,
-    photo: MultipartFile,
-    caption: String? = null,
-    parseMode: ParseMode? = null,
-    disableNotification: Boolean = false,
-    replyToMessageId: MessageIdentifier? = null,
-    replyMarkup: KeyboardMarkup? = null
-) = execute(
-    SendPhoto(
-        chatId,
-        photo,
-        caption,
-        parseMode,
-        disableNotification,
-        replyToMessageId,
-        replyMarkup
-    )
-)
-
-suspend fun RequestsExecutor.sendPhoto(
-    chatId: ChatIdentifier,
-    file: PhotoSize,
-    caption: String? = null,
-    parseMode: ParseMode? = null,
-    disableNotification: Boolean = false,
-    replyToMessageId: MessageIdentifier? = null,
-    replyMarkup: KeyboardMarkup? = null
-) = sendPhoto(
-    chatId, file.fileId, caption, parseMode, disableNotification, replyToMessageId, replyMarkup
-)
-
-suspend fun RequestsExecutor.sendPhoto(
-    chatId: ChatIdentifier,
-    photo: Photo,
-    caption: String? = null,
-    parseMode: ParseMode? = null,
-    disableNotification: Boolean = false,
-    replyToMessageId: MessageIdentifier? = null,
-    replyMarkup: KeyboardMarkup? = null
-) = sendPhoto(
-    chatId, photo.biggest() ?: throw IllegalArgumentException("Photo $photo is empty"), caption, parseMode, disableNotification, replyToMessageId, replyMarkup
-)
-
-suspend fun RequestsExecutor.sendPhoto(
-    chat: Chat,
-    fileId: FileId,
-    caption: String? = null,
-    parseMode: ParseMode? = null,
-    disableNotification: Boolean = false,
-    replyToMessageId: MessageIdentifier? = null,
-    replyMarkup: KeyboardMarkup? = null
-) = sendPhoto(
-    chat.id,
-    fileId,
-    caption,
-    parseMode,
-    disableNotification,
-    replyToMessageId,
-    replyMarkup
-)
-
-suspend fun RequestsExecutor.sendPhoto(
-    chat: Chat,
-    file: PhotoSize,
-    caption: String? = null,
-    parseMode: ParseMode? = null,
-    disableNotification: Boolean = false,
-    replyToMessageId: MessageIdentifier? = null,
-    replyMarkup: KeyboardMarkup? = null
-) = sendPhoto(
-    chat.id, file.fileId, caption, parseMode, disableNotification, replyToMessageId, replyMarkup
-)
-
-suspend fun RequestsExecutor.sendPhoto(
-    chat: Chat,
-    photo: Photo,
-    caption: String? = null,
-    parseMode: ParseMode? = null,
-    disableNotification: Boolean = false,
-    replyToMessageId: MessageIdentifier? = null,
-    replyMarkup: KeyboardMarkup? = null
-) = sendPhoto(
-    chat.id, photo.biggest() ?: throw IllegalArgumentException("Photo $photo is empty"), caption, parseMode, disableNotification, replyToMessageId, replyMarkup
-)
-
-suspend fun RequestsExecutor.sendPhoto(
-    chat: Chat,
-    fileId: MultipartFile,
-    caption: String? = null,
-    parseMode: ParseMode? = null,
-    disableNotification: Boolean = false,
-    replyToMessageId: MessageIdentifier? = null,
-    replyMarkup: KeyboardMarkup? = null
-) = sendPhoto(
-    chat.id,
-    fileId,
-    caption,
-    parseMode,
-    disableNotification,
-    replyToMessageId,
-    replyMarkup
 )