+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * `TelegramBotAPI`:
     * All suspend functions for `RequestsExecutor` was removed (due to replacement into
     [TelegramBotAPI extensions project](TelegramBotAPI-extensions-api/README.md))
+* `TelegramBotAPI-extensions-api`:
+    * Most part of sending media messages functions was removed and replaced with their `InputFile` args analogs
+
 
 ## 0.23.0 TelegramBotAPI 4.6